% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-recruitment.R
\name{bb_fit_recruitment_ml}
\alias{bb_fit_recruitment_ml}
\title{Fit Recruitment Model with Maximum Likelihood}
\usage{
bb_fit_recruitment_ml(
  data,
  adult_female_proportion = 0.65,
  sex_ratio = 0.5,
  min_random_year = 5,
  year_trend = FALSE,
  year_start = 4L,
  inits = NULL,
  quiet = FALSE
)
}
\arguments{
\item{data}{The data.frame.}

\item{adult_female_proportion}{A number between 0 and 1 of the expected proportion of adults that are female.
If NULL, the proportion is estimated from the data (i.e., \code{Cows ~ Binomial(adult_female_proportion, Cows + Bulls)}) and a prior of dbeta(65, 35) is used.
This prior can be changed via the \code{priors} argument.}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth.
This proportion is applied to yearlings.}

\item{min_random_year}{A whole number of the minimum number of years required to fit year as a random effect (as opposed to a fixed effect).}

\item{year_trend}{A flag indicating whether to fit a year trend effect.
Year trend cannot be fit if there is also a fixed year effect (as opposed to random effect).}

\item{year_start}{A whole number between 1 and 12 indicating the start of the caribou (i.e., biological) year. By default, April is set as the start of the caribou year.}

\item{inits}{A named vector of the parameter initial values.
Any missing values are assigned a default value of 0.
If NULL, all parameters are assigned a default value of 0.}

\item{quiet}{A flag indicating whether to suppress messages and progress bars.}
}
\value{
A list of the Nimble model object and Maximum Likelihood output with estimates and standard errors on the transformed scale.
}
\description{
Fit recruitment model with Maximum Likelihood using Nimble Laplace Approximation.
}
\details{
If the number of years is > \code{min_random_year}, a fixed-effects model is fit.
Otherwise, a mixed-effects model is fit with random intercept for each year.
If \code{year_trend} is TRUE and the number of years is > \code{min_random_year}, the model
will be fit with year as a continuous effect (i.e. trend) and no fixed effect of year.
If \code{year_trend} is TRUE and the number of years is <= \code{min_random_year}, the model
will be fit with year as a continuous effect and a random intercept for each year.

Year effect can be excluded with \code{exclude_year}. This can be useful if the ML model is failing to converge.

The start month of the Caribou year can be adjusted with \code{year_start}.
}
\examples{
if (interactive()) {
  fit <- bb_fit_recruitment_ml(bboudata::bbourecruit_a)
}
}
\seealso{
Other model: 
\code{\link{bb_fit_recruitment}()},
\code{\link{bb_fit_survival}()},
\code{\link{bb_fit_survival_ml}()}
}
\concept{model}
