% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-month.R
\name{bb_plot_month}
\alias{bb_plot_month}
\alias{bb_plot_month.data.frame}
\alias{bb_plot_month.bboufit_survival}
\title{Plot Month}
\usage{
bb_plot_month(x, ...)

\method{bb_plot_month}{data.frame}(x, ...)

\method{bb_plot_month}{bboufit_survival}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plots month estimates with credible limits.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_month(data.frame)}: Plot monthly estimate for a data frame.

\item \code{bb_plot_month(bboufit_survival)}: Plot monthly estimates for a bboufit_survival object.

}}
