% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-month-survival.R
\name{bb_plot_month_survival}
\alias{bb_plot_month_survival}
\alias{bb_plot_month_survival.data.frame}
\alias{bb_plot_month_survival.bboufit_survival}
\title{Plot Monthly Survival}
\usage{
bb_plot_month_survival(x, ...)

\method{bb_plot_month_survival}{data.frame}(x, ...)

\method{bb_plot_month_survival}{bboufit_survival}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plots monthly survival estimates with credible limits.
Estimates represent annual survival if a given month lasted the entire year.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_month_survival(data.frame)}: Plot monthly survival estimate for a data frame.

\item \code{bb_plot_month_survival(bboufit_survival)}: Plot monthly survival estimates for a bboufit_survival object.

}}
