% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year.R
\name{bb_plot_year}
\alias{bb_plot_year}
\alias{bb_plot_year.data.frame}
\alias{bb_plot_year.bboufit}
\alias{bb_plot_year.bboufit_ml}
\title{Plot Year}
\usage{
bb_plot_year(x, ...)

\method{bb_plot_year}{data.frame}(x, ...)

\method{bb_plot_year}{bboufit}(x, conf_level = 0.95, estimate = median, ...)

\method{bb_plot_year}{bboufit_ml}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plots annual estimates with credible limits.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_year(data.frame)}: Plot annual estimate for a data frame.

\item \code{bb_plot_year(bboufit)}: Plot annual estimates for a bboufit object.

\item \code{bb_plot_year(bboufit_ml)}: Plot annual estimates for a bboufit_ml object.

}}
