% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-calf-cow-ratio.R
\name{bb_plot_year_calf_cow_ratio}
\alias{bb_plot_year_calf_cow_ratio}
\alias{bb_plot_year_calf_cow_ratio.data.frame}
\alias{bb_plot_year_calf_cow_ratio.bboufit_recruitment}
\title{Plot Year Calf-Cow Ratio}
\usage{
bb_plot_year_calf_cow_ratio(x, ...)

\method{bb_plot_year_calf_cow_ratio}{data.frame}(x, ...)

\method{bb_plot_year_calf_cow_ratio}{bboufit_recruitment}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plot annual calf-cow ratio estimates with credible limits.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_year_calf_cow_ratio(data.frame)}: Plot annual recruitment estimate for a data frame.

\item \code{bb_plot_year_calf_cow_ratio(bboufit_recruitment)}: Plot annual calf-cow ratio estimates for a bboufit_recruitment object.

}}
