% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-growth.R
\name{bb_plot_year_growth}
\alias{bb_plot_year_growth}
\title{Plot Year Population Growth}
\usage{
bb_plot_year_growth(x)
}
\arguments{
\item{x}{A data frame of the lambda estimates (output of \code{\link[=bb_predict_growth]{bb_predict_growth()}}).}
}
\description{
Plots annual population growth with credible limits.
}
