% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-growth.R
\name{bb_plot_year_population_change}
\alias{bb_plot_year_population_change}
\title{Plot Year Population Change}
\usage{
bb_plot_year_population_change(x)
}
\arguments{
\item{x}{A data frame of the population change estimates (output of \code{\link[=bb_predict_population_change]{bb_predict_population_change()}}).}
}
\description{
Plots annual population change (\%) with credible limits.
}
