% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-recruitment.R
\name{bb_plot_year_recruitment}
\alias{bb_plot_year_recruitment}
\alias{bb_plot_year_recruitment.data.frame}
\alias{bb_plot_year_recruitment.bboufit_recruitment}
\title{Plot Year Recruitment}
\usage{
bb_plot_year_recruitment(x, ...)

\method{bb_plot_year_recruitment}{data.frame}(x, ...)

\method{bb_plot_year_recruitment}{bboufit_recruitment}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plot annual recruitment estimates with credible limits.
Recruitment is adjusted following DeCesare et al. (2012) methods.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_year_recruitment(data.frame)}: Plot annual recruitment estimate for a data frame.

\item \code{bb_plot_year_recruitment(bboufit_recruitment)}: Plot annual recruitment estimates for a bboufit_recruitment object.

}}
\references{
DeCesare, Nicholas J., Mark Hebblewhite, Mark Bradley, Kirby G. Smith,
David Hervieux, and Lalenia Neufeld. 2012 “Estimating Ungulate Recruitment
and Growth Rates Using Age Ratios.” The Journal of Wildlife Management
76 (1): 144–53 https://doi.org/10.1002/jwmg.244.
}
