% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-survival.R
\name{bb_plot_year_survival}
\alias{bb_plot_year_survival}
\alias{bb_plot_year_survival.data.frame}
\alias{bb_plot_year_survival.bboufit_survival}
\title{Plot Year Survival}
\usage{
bb_plot_year_survival(x, ...)

\method{bb_plot_year_survival}{data.frame}(x, ...)

\method{bb_plot_year_survival}{bboufit_survival}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plots annual survival estimates with credible limits.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_year_survival(data.frame)}: Plot annual survival estimate for a data frame.

\item \code{bb_plot_year_survival(bboufit_survival)}: Plot annual survival estimates for a bboufit_survival object.

}}
