% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-trend-recruitment.R
\name{bb_plot_year_trend_recruitment}
\alias{bb_plot_year_trend_recruitment}
\alias{bb_plot_year_trend_recruitment.data.frame}
\alias{bb_plot_year_trend_recruitment.bboufit_recruitment}
\title{Plot Annual Recruitment Trend}
\usage{
bb_plot_year_trend_recruitment(x, ...)

\method{bb_plot_year_trend_recruitment}{data.frame}(x, ...)

\method{bb_plot_year_trend_recruitment}{bboufit_recruitment}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plots annual recruitment estimates as trend line with credible limits.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_year_trend_recruitment(data.frame)}: Plot annual recruitment estimate as trend line for a data frame.

\item \code{bb_plot_year_trend_recruitment(bboufit_recruitment)}: Plot annual estimates as trend line for a bboufit_recruitment object.

}}
