% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-trend-survival.R
\name{bb_plot_year_trend_survival}
\alias{bb_plot_year_trend_survival}
\alias{bb_plot_year_trend_survival.data.frame}
\alias{bb_plot_year_trend_survival.bboufit_survival}
\title{Plot Annual Survival Trend}
\usage{
bb_plot_year_trend_survival(x, ...)

\method{bb_plot_year_trend_survival}{data.frame}(x, ...)

\method{bb_plot_year_trend_survival}{bboufit_survival}(x, conf_level = 0.95, estimate = median, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\description{
Plots annual survival estimates as trend line with credible limits.
}
\section{Methods (by class)}{
\itemize{
\item \code{bb_plot_year_trend_survival(data.frame)}: Plot annual survival estimate as trend line for a data frame.

\item \code{bb_plot_year_trend_survival(bboufit_survival)}: Plot annual estimates as trend line for a bboufit_survival object.

}}
