% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{bb_predict_calf_cow_ratio_samples}
\alias{bb_predict_calf_cow_ratio_samples}
\title{Predict Calf-Cow Ratio Samples}
\usage{
bb_predict_calf_cow_ratio_samples(recruitment, year = TRUE)
}
\arguments{
\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{year}{A flag indicating whether to predict by year.}
}
\value{
A 'mcmcarray' object with the modified MCMC samples.
}
\description{
Predict calves per adult female by year.
If year is FALSE, predictions are made for a 'typical' year.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_calf_cow_ratio_trend_samples}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_growth_samples}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_population_change_samples}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_samples}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_recruitment_trend_samples}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_samples}()},
\code{\link{bb_predict_survival_trend}()},
\code{\link{bb_predict_survival_trend_samples}()}
}
\concept{analysis}
