% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-trend.R
\name{bb_predict_calf_cow_ratio_trend}
\alias{bb_predict_calf_cow_ratio_trend}
\title{Predict Calf-Cow Ratio Trend}
\usage{
bb_predict_calf_cow_ratio_trend(
  recruitment,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 5
)
}
\arguments{
\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}
}
\value{
A tibble of the predicted estimates.
}
\description{
Predict calves per adult female by year as trend line.
Recruitment fit object provided must be created with \code{year_trend = TRUE}.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_trend}()}
}
\concept{analysis}
