% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-growth.R
\name{bb_predict_growth}
\alias{bb_predict_growth}
\alias{bb_predict_lambda}
\title{Predict Population Growth Lambda}
\usage{
bb_predict_growth(
  survival,
  recruitment,
  sex_ratio = 0.5,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3
)

bb_predict_lambda(
  survival,
  recruitment,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3
)
}
\arguments{
\item{survival}{An object of class 'bboufit_survival' (output of \code{\link[=bb_fit_survival]{bb_fit_survival()}}).}

\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}
}
\value{
A tibble of the lambda estimates with upper and lower credible intervals.
}
\description{
Predicts population growth (lambda) from survival and recruitment fit objects using the Hatter-Bergerud equation
(Hatter and Bergerud, 1991).
}
\section{Functions}{
\itemize{
\item \code{bb_predict_lambda()}: Deprecated for \code{bb_predict_growth()} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

}}
\examples{
if (interactive()) {
  survival <- bb_fit_survival(bboudata::bbousurv_a)
  recruitment <- bb_fit_recruitment(bboudata::bbourecruit_a)
  growth <- bb_predict_growth(survival, recruitment)
}
}
\references{
Hatter, Ian, and Wendy Bergerud. 1991. “Moose Recruitment, Adult
Mortality and Rate of Change” 27: 65–73.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_trend}()}
}
\concept{analysis}
