% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-growth.R
\name{bb_predict_population_change}
\alias{bb_predict_population_change}
\title{Predict Population Change}
\usage{
bb_predict_population_change(
  survival,
  recruitment,
  sex_ratio = 0.5,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3
)
}
\arguments{
\item{survival}{An object of class 'bboufit_survival' (output of \code{\link[=bb_fit_survival]{bb_fit_survival()}}).}

\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}
}
\value{
A tibble of the population change estimates with upper and lower credible intervals.
}
\description{
Predicts population change (\%) from survival and recruitment fit objects.
Population change is the cumulative product of population growth rate (i.e., output of \code{\link[=bb_predict_growth]{bb_predict_growth()}})
}
\examples{
if (interactive()) {
  survival <- bb_fit_survival(bboudata::bbousurv_a)
  recruitment <- bb_fit_recruitment(bboudata::bbourecruit_a)
  change <- bb_predict_population_change(survival, recruitment)
}
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_trend}()}
}
\concept{analysis}
