% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-growth.R
\name{bb_predict_population_change_samples}
\alias{bb_predict_population_change_samples}
\title{Predict Population Change Samples}
\usage{
bb_predict_population_change_samples(
  survival,
  recruitment,
  sex_ratio = deprecated()
)
}
\arguments{
\item{survival}{An object of class 'bboufit_survival' (output of \code{\link[=bb_fit_survival]{bb_fit_survival()}}).}

\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth. Deprecated for \code{bb_fit_recruitment(sex_ratio)}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
A list with elements:
\itemize{
\item \code{samples}: An \code{mcmcarray} with the MCMC samples.
\item \code{data}: A \code{data.frame} containing the associated prediction data.
}
}
\description{
Predicts population change (\%) from survival and recruitment fit objects.
Population change is the cumulative product of population growth rate (i.e., output of \code{\link[=bb_predict_growth]{bb_predict_growth()}})
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_samples}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_calf_cow_ratio_trend_samples}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_growth_samples}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_samples}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_recruitment_trend_samples}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_samples}()},
\code{\link{bb_predict_survival_trend}()},
\code{\link{bb_predict_survival_trend_samples}()}
}
\concept{analysis}
