% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{bb_predict_recruitment}
\alias{bb_predict_recruitment}
\title{Predict Recruitment}
\usage{
bb_predict_recruitment(
  recruitment,
  year = TRUE,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3
)
}
\arguments{
\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{year}{A flag indicating whether to predict by year.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}
}
\value{
A tibble of the predicted estimates.
}
\description{
Predict recruitment by year.
If year is FALSE, predictions are made for a 'typical' year.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_trend}()}
}
\concept{analysis}
