% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{bb_predict_recruitment_samples}
\alias{bb_predict_recruitment_samples}
\title{Predict Recruitment Samples}
\usage{
bb_predict_recruitment_samples(recruitment, year = TRUE, sex_ratio = 0.5)
}
\arguments{
\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{year}{A flag indicating whether to predict by year.}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth.}
}
\value{
A 'mcmcarray' object with the modified MCMC samples.
}
\description{
Predict adjusted recruitment by year using DeCesare et al. (2012) methods.
If year is FALSE, predictions are made for a 'typical' year.
See \code{\link[=bb_predict_calf_cow_ratio]{bb_predict_calf_cow_ratio()}} for unadjusted recruitment.
}
\references{
DeCesare, Nicholas J., Mark Hebblewhite, Mark Bradley, Kirby G. Smith,
David Hervieux, and Lalenia Neufeld. 2012 “Estimating Ungulate Recruitment
and Growth Rates Using Age Ratios.” The Journal of Wildlife Management
76 (1): 144–53 https://doi.org/10.1002/jwmg.244.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_samples}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_calf_cow_ratio_trend_samples}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_growth_samples}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_population_change_samples}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_recruitment_trend_samples}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_samples}()},
\code{\link{bb_predict_survival_trend}()},
\code{\link{bb_predict_survival_trend_samples}()}
}
\concept{analysis}
