% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-trend.R
\name{bb_predict_recruitment_trend_samples}
\alias{bb_predict_recruitment_trend_samples}
\title{Predict Recruitment Trend Samples}
\usage{
bb_predict_recruitment_trend_samples(recruitment, sex_ratio = deprecated())
}
\arguments{
\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth. Deprecated for \code{bb_fit_recruitment(sex_ratio)}. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
A 'mcmcarray' object containing the modified MCMC samples.
}
\description{
Predict recruitment by year as trend line.
Recruitment fit object provided must be created with \code{year_trend = TRUE}.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_samples}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_calf_cow_ratio_trend_samples}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_growth_samples}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_population_change_samples}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_samples}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_samples}()},
\code{\link{bb_predict_survival_trend}()},
\code{\link{bb_predict_survival_trend_samples}()}
}
\concept{analysis}
