% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{bb_predict_survival}
\alias{bb_predict_survival}
\title{Predict Survival}
\usage{
bb_predict_survival(
  survival,
  year = TRUE,
  month = FALSE,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3
)
}
\arguments{
\item{survival}{An object of class 'bboufit_survival' (output of \code{\link[=bb_fit_survival]{bb_fit_survival()}}).}

\item{year}{A flag indicating whether to predict by year.}

\item{month}{A flag indicating whether to predict by month.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}
}
\value{
A tibble of the predicted estimates.
}
\description{
Predict survival by year and/or month.
If year and month are FALSE, predictions are made for a 'typical' year and month.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_survival_trend}()}
}
\concept{analysis}
