% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{bb_predict_survival_samples}
\alias{bb_predict_survival_samples}
\title{Predict Survival Samples}
\usage{
bb_predict_survival_samples(survival, year = TRUE, month = FALSE)
}
\arguments{
\item{survival}{An object of class 'bboufit_survival' (output of \code{\link[=bb_fit_survival]{bb_fit_survival()}}).}

\item{year}{A flag indicating whether to predict by year.}

\item{month}{A flag indicating whether to predict by month.}
}
\value{
A list with elements:
\itemize{
\item \code{samples}: An \code{mcmcarray} with the MCMC samples.
\item \code{data}: A \code{data.frame} containing the associated prediction data.
}
}
\description{
Predict survival by year and/or month.
If year and month are FALSE, predictions are made for a 'typical' year and month.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_samples}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_calf_cow_ratio_trend_samples}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_growth_samples}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_population_change_samples}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_samples}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_recruitment_trend_samples}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_trend}()},
\code{\link{bb_predict_survival_trend_samples}()}
}
\concept{analysis}
