% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-trend.R
\name{bb_predict_survival_trend_samples}
\alias{bb_predict_survival_trend_samples}
\title{Predict Survival Trend Samples}
\usage{
bb_predict_survival_trend_samples(survival)
}
\arguments{
\item{survival}{An object of class 'bboufit_survival' (output of \code{\link[=bb_fit_survival]{bb_fit_survival()}}).}
}
\value{
A list with elements:
\itemize{
\item \code{samples}: An \code{mcmcarray} with the MCMC samples.
\item \code{data}: A \code{data.frame} containing the associated prediction data.
}
}
\description{
Predict survival by year as trend line.
Survival fit object provided must be created with \code{year_trend = TRUE}.
}
\seealso{
Other analysis: 
\code{\link{bb_predict_calf_cow_ratio}()},
\code{\link{bb_predict_calf_cow_ratio_samples}()},
\code{\link{bb_predict_calf_cow_ratio_trend}()},
\code{\link{bb_predict_calf_cow_ratio_trend_samples}()},
\code{\link{bb_predict_growth}()},
\code{\link{bb_predict_growth_samples}()},
\code{\link{bb_predict_population_change}()},
\code{\link{bb_predict_population_change_samples}()},
\code{\link{bb_predict_recruitment}()},
\code{\link{bb_predict_recruitment_samples}()},
\code{\link{bb_predict_recruitment_trend}()},
\code{\link{bb_predict_recruitment_trend_samples}()},
\code{\link{bb_predict_survival}()},
\code{\link{bb_predict_survival_samples}()},
\code{\link{bb_predict_survival_trend}()}
}
\concept{analysis}
