% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bb_priors_survival}
\alias{bb_priors_survival}
\title{Survival model default priors}
\usage{
bb_priors_survival()
}
\value{
A named vector.
}
\description{
Prior distribution parameters and default values for survival model parameters.
}
\details{
Intercept

\code{b0 ~ Normal(mu = b0_mu, sd = b0_sd)}

Year Trend

\code{bYear ~ Normal(mu = bYear_mu, sd = bYear_sd)}

Year fixed effect

\code{bAnnual ~ Normal(mu = 0, sd = bAnnual_sd)}

Standard deviation of annual random effect

\code{sAnnual ~ Exponential(rate = sAnnual_rate)}

Standard deviation of month random effect

\code{sMonth ~ Exponential(rate = sMonth_rate)}
}
\examples{
bb_priors_survival()
}
