% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboutools-package.R
\docType{package}
\name{bboutools-package}
\alias{bboutools}
\alias{bboutools-package}
\title{bboutools: Boreal Caribou Survival, Recruitment and Population Growth}
\description{
Estimates annual survival, recruitment and population growth for boreal caribou populations using Bayesian and Maximum Likelihood models with fixed and random effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/bboutools}
  \item \url{https://poissonconsulting.github.io/bboutools/}
  \item Report bugs at \url{https://github.com/poissonconsulting/bboutools/issues}
}

}
\author{
\strong{Maintainer}: Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
  \item John Boulanger (\href{https://orcid.org/0000-0001-8222-1445}{ORCID})
  \item Ayla Pearson \email{ayla@poissonconsulting.ca}
}

Other contributors:
\itemize{
  \item Province of Alberta [copyright holder]
}

}
\keyword{internal}
