% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.bboufit}
\alias{coef.bboufit}
\title{Get Tidy Tibble from bboufit Object.}
\usage{
\method{coef}{bboufit}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused parameters.}
}
\description{
A wrapper on \code{\link[=tidy.bboufit]{tidy.bboufit()}}.
}
\examples{
if (interactive()) {
  fit <- bb_fit_recruitment(bboudata::bbourecruit_a)
  coef(fit)
}
}
\seealso{
\code{\link[=tidy.bboufit]{tidy.bboufit()}}
}
