% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.bboufit_ml}
\alias{coef.bboufit_ml}
\title{Get Tidy Tibble from bboufit_ml Object.}
\usage{
\method{coef}{bboufit_ml}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused parameters.}
}
\description{
A wrapper on \code{\link[=tidy.bboufit_ml]{tidy.bboufit_ml()}}.
}
\examples{
if (interactive()) {
  fit <- bb_fit_recruitment_ml(bboudata::bbourecruit_a)
  coef(fit)
}
}
\seealso{
\code{\link[=tidy.bboufit_ml]{tidy.bboufit_ml()}}
}
