% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{converged.bboufit}
\alias{converged.bboufit}
\title{Get Convergence of bboufit Object}
\usage{
\method{converged}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A flag indicating convergence.
}
\description{
Get Convergence of bboufit Object
}
