% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{converged.bboufit_ml}
\alias{converged.bboufit_ml}
\title{Get Convergence of bboufit_ml Object}
\usage{
\method{converged}{bboufit_ml}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A flag indicating convergence.
}
\description{
Successful convergence indicates that no convergence warnings were produced
by optim and all standard errors could be estimated.
}
