% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{esr.bboufit}
\alias{esr.bboufit}
\title{Get Effective Sample Rate of bboufit Object}
\usage{
\method{esr}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of chains.
}
\description{
Get Effective Sample Rate of bboufit Object
}
