% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{estimates.bboufit}
\alias{estimates.bboufit}
\title{Estimates for bboufit Object}
\usage{
\method{estimates}{bboufit}(x, term = NULL, ...)
}
\arguments{
\item{x}{The object.}

\item{term}{A string of the term name.}

\item{...}{Unused parameters.}
}
\value{
A named list of the estimates.
}
\description{
Gets a named list of the estimated values by term.
}
\examples{
if (interactive()) {
  fit <- bb_fit_survival(bboudata::bbousurv_a)
  estimates(fit)
}
}
\seealso{
\code{\link[=tidy.bboufit]{tidy.bboufit()}}
}
