% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{estimates.bboufit_ml}
\alias{estimates.bboufit_ml}
\title{Estimates for bboufit_ml Object}
\usage{
\method{estimates}{bboufit_ml}(x, term = NULL, original_scale = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{term}{A string of the term name.}

\item{original_scale}{A flag indicating whether to return the estimates in the original scale.}

\item{...}{Unused parameters.}
}
\value{
A named list of the estimates.
}
\description{
Gets a named list of the estimated values by term.
}
\examples{
if (interactive()) {
  fit <- bb_fit_survival_ml(bboudata::bbousurv_a)
  estimates(fit)
}
}
\seealso{
\code{\link[=tidy.bboufit]{tidy.bboufit()}}
}
