% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.bboufit_ml}
\alias{glance.bboufit_ml}
\title{Get a Glance Summary of bboufit_ml Object}
\usage{
\method{glance}{bboufit_ml}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A tibble of the glance summary.
}
\description{
Get a tibble of a one-row summary of the model fit.
}
\examples{
if (interactive()) {
  fit <- bb_fit_survival_ml(bboudata::bbousurv_a)
  glance(fit)
}
}
\seealso{
Other generics: 
\code{\link{augment.bboufit}()},
\code{\link{augment.bboufit_ml}()},
\code{\link{glance.bboufit}()},
\code{\link{tidy.bboufit}()},
\code{\link{tidy.bboufit_ml}()}
}
\concept{generics}
