% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-code.R
\name{model_code}
\alias{model_code}
\alias{model_code.bboufit}
\alias{model_code.bboufit_ml}
\title{Get Model Code}
\usage{
model_code(x, ...)

\method{model_code}{bboufit}(x, ...)

\method{model_code}{bboufit_ml}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\description{
Get code from Nimble model.
}
\section{Methods (by class)}{
\itemize{
\item \code{model_code(bboufit)}: Get model code from bboufit object.

\item \code{model_code(bboufit_ml)}: Get model code from bboufit_ml object.

}}
