% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-survival.R
\name{model_survival}
\alias{model_survival}
\title{Build Nimble survival model.}
\usage{
model_survival(
  data,
  year_random = TRUE,
  year_trend = FALSE,
  priors = NULL,
  exclude_year = FALSE,
  build_derivs = TRUE
)
}
\arguments{
\item{data}{The data.frame.}

\item{year_random}{A flag indicating whether to include year random effect. If FALSE, year is fitted as a fixed effect.}

\item{year_trend}{A flag indicating whether to fit a year trend effect.
Year trend cannot be fit if there is also a fixed year effect (as opposed to random effect).}

\item{priors}{A named vector of the parameter prior distribution values.
Any missing values are assigned their default values in \code{priors_survival()} and \code{priors_recruitment()}.
If NULL, all parameters are assigned their default priors.}

\item{exclude_year}{A flag indicating whether to exclude the year effect.}

\item{build_derivs}{A flag indicating whether to build derivatives Laplace approximation.}
}
\description{
This is for use by developers.
}
\keyword{Internal}
