% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{nchains.bboufit}
\alias{nchains.bboufit}
\title{Get Number of Chains from bboufit Object}
\usage{
\method{nchains}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of chains.
}
\description{
Get Number of Chains from bboufit Object
}
