% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{niters.bboufit}
\alias{niters.bboufit}
\title{Get Number of Iterations from bboufit Object}
\usage{
\method{niters}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of iterations.
}
\description{
Get Number of Iterations from bboufit Object
}
