% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{npars.bboufit}
\alias{npars.bboufit}
\title{Get Number of Parameters from bboufit Object}
\usage{
\method{npars}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of parameters.
}
\description{
Get Number of Parameters from bboufit Object
}
