% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{npars.bboufit_ml}
\alias{npars.bboufit_ml}
\title{Get Number of Parameters from bboufit_ml Object}
\usage{
\method{npars}{bboufit_ml}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of parameters.
}
\description{
Get Number of Parameters from bboufit_ml Object
}
