% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{nterms.bboufit}
\alias{nterms.bboufit}
\title{Get Number of Terms from bboufit Object}
\usage{
\method{nterms}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of terms.
}
\description{
Get Number of Terms from bboufit Object
}
