% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for bboutools Functions}
\arguments{
\item{...}{Unused parameters.}

\item{data}{The data.frame.}

\item{x}{The object.}

\item{object}{The object.}

\item{by}{A string of the column indicating group to derive predictions at.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{include_random_effects}{A flag indicating whether to include random effects in coefficient table. Standard deviation estimates will always be included.}

\item{estimate}{A function to calculate the estimate.}

\item{min_random_year}{A whole number of the minimum number of years required to fit year as a random effect (as opposed to a fixed effect).}

\item{year_trend}{A flag indicating whether to fit a year trend effect.
Year trend cannot be fit if there is also a fixed year effect (as opposed to random effect).}

\item{include_uncertain_morts}{A flag indicating whether to include uncertain mortalities in total mortalities.}

\item{yearling_female_proportion}{A number between 0 and 1 of the expected proportion of yearlings that are female.}

\item{adult_female_proportion}{A number between 0 and 1 of the expected proportion of adults that are female.
If NULL, a prior of dbeta(65, 35) is used. This prior can also be changed via the \code{priors} argument.}

\item{year}{A flag indicating whether to predict by year.}

\item{month}{A flag indicating whether to predict by month.}

\item{nthin}{A whole number of the thinning rate.}

\item{niters}{A whole number of the number of iterations per chain after thinning and burn-in.}

\item{priors}{A named vector of the parameter prior distribution values.
Any missing values are assigned their default values in \code{priors_survival()} and \code{priors_recruitment()}.
If NULL, all parameters are assigned their default priors.}

\item{inits}{A named vector of the parameter initial values.
Any missing values are assigned a default value of 0.
If NULL, all parameters are assigned a default value of 0.}

\item{survival}{An object of class 'bboufit_survival' (output of \code{\link[=bb_fit_survival]{bb_fit_survival()}}).}

\item{recruitment}{An object of class 'bboufit_recruitment' (output of \code{\link[=bb_fit_recruitment]{bb_fit_recruitment()}})}

\item{quiet}{A flag indicating whether to suppress messages and progress bars.}

\item{term}{A string of the term name.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}

\item{original_scale}{A flag indicating whether to return the estimates in the original scale.}

\item{exclude_year}{A flag indicating whether to exclude the year effect.}

\item{year_random}{A flag indicating whether to include year random effect. If FALSE, year is fitted as a fixed effect.}

\item{year_start}{A whole number between 1 and 12 indicating the start of the caribou year. By default, April is set as the start of the biological year.}

\item{params}{A named list of the parameter names and values to simulate.}
}
\description{
Parameter Descriptions for bboutools Functions
}
\keyword{internal}
