% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{pars.bboufit}
\alias{pars.bboufit}
\title{Get Parameters from bboufit Object}
\usage{
\method{pars}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A vector of the parameter names.
}
\description{
Get Parameters from bboufit Object
}
