% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.bboufit_recruitment}
\alias{predict.bboufit_recruitment}
\title{Predict Recruitment}
\usage{
\method{predict}{bboufit_recruitment}(
  object,
  year = TRUE,
  sex_ratio = 0.5,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{year}{A flag indicating whether to predict by year.}

\item{sex_ratio}{A number between 0 and 1 of the proportion of females at birth.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}

\item{...}{Unused parameters.}
}
\description{
A wrapper on \code{\link[=bb_predict_recruitment]{bb_predict_recruitment()}}.
}
\seealso{
\code{\link[=bb_predict_recruitment]{bb_predict_recruitment()}}
}
