% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.bboufit_survival}
\alias{predict.bboufit_survival}
\title{Predict Survival}
\usage{
\method{predict}{bboufit_survival}(
  object,
  year = TRUE,
  month = FALSE,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{year}{A flag indicating whether to predict by year.}

\item{month}{A flag indicating whether to predict by month.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}

\item{...}{Unused parameters.}
}
\description{
A wrapper on \code{\link[=bb_predict_survival]{bb_predict_survival()}}.
}
\seealso{
\code{\link[=bb_predict_survival]{bb_predict_survival()}}
}
