% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{rhat.bboufit}
\alias{rhat.bboufit}
\title{Get Rhat of bboufit Object}
\usage{
\method{rhat}{bboufit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of rhat value.
}
\description{
Get Rhat of bboufit Object
}
