% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples}
\alias{samples}
\alias{samples.bboufit}
\alias{samples.bboufit_ml}
\title{Get MCMC samples}
\usage{
samples(x)

\method{samples}{bboufit}(x)

\method{samples}{bboufit_ml}(x)
}
\arguments{
\item{x}{The object.}
}
\description{
Get MCMC samples from Nimble model.
}
\section{Methods (by class)}{
\itemize{
\item \code{samples(bboufit)}: Get MCMC samples from bboufit object.

\item \code{samples(bboufit_ml)}: Create MCMC samples (1 iteration, 1 chain) from bboufit_ml object.

}}
