% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.bboufit}
\alias{tidy.bboufit}
\title{Get Tidy Tibble from bboufit Object.}
\usage{
\method{tidy}{bboufit}(
  x,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3,
  include_random_effects = TRUE,
  ...
)
}
\arguments{
\item{x}{The object.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}

\item{include_random_effects}{A flag indicating whether to include random effects in coefficient table. Standard deviation estimates will always be included.}

\item{...}{Unused parameters.}
}
\value{
A tibble of the tidy coefficient summary.
}
\description{
Get a tidy tibble of the coefficient estimates and
confidence intervals from Bayesian model fit.
}
\examples{
if (interactive()) {
  fit <- bb_fit_survival(bboudata::bbousurv_a)
  tidy(fit)
}
}
\seealso{
\code{\link[=coef.bboufit]{coef.bboufit()}}

Other generics: 
\code{\link{augment.bboufit}()},
\code{\link{augment.bboufit_ml}()},
\code{\link{glance.bboufit}()},
\code{\link{glance.bboufit_ml}()},
\code{\link{tidy.bboufit_ml}()}
}
\concept{generics}
