% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.bboufit_ml}
\alias{tidy.bboufit_ml}
\title{Get Tidy Tibble from bboufit_ml Object.}
\usage{
\method{tidy}{bboufit_ml}(x, conf_level = 0.95, sig_fig = 3, include_random_effects = TRUE, ...)
}
\arguments{
\item{x}{The object.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{sig_fig}{A whole number of the significant figures to round estimates by.}

\item{include_random_effects}{A flag indicating whether to include random effects in coefficient table. Standard deviation estimates will always be included.}

\item{...}{Unused parameters.}
}
\value{
A tibble of the tidy coefficient summary.
}
\description{
Get a tidy tibble of the coefficient estimates and
confidence intervals from Maximum Likelihood model fit.
}
\examples{
if (interactive()) {
  fit <- bb_fit_survival_ml(bboudata::bbousurv_a)
  tidy(fit)
}
}
\seealso{
\code{\link[=coef.bboufit_ml]{coef.bboufit_ml()}}

Other generics: 
\code{\link{augment.bboufit}()},
\code{\link{augment.bboufit_ml}()},
\code{\link{glance.bboufit}()},
\code{\link{glance.bboufit_ml}()},
\code{\link{tidy.bboufit}()}
}
\concept{generics}
