% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-recruitment.R
\name{model_recruitment}
\alias{model_recruitment}
\title{Build Nimble recruitment model.}
\usage{
model_recruitment(
  data,
  year_random = TRUE,
  year_trend = TRUE,
  adult_female_proportion = 0.65,
  yearling_female_proportion = 0.5,
  demographic_stochasticity = TRUE,
  priors = NULL
)
}
\arguments{
\item{data}{The data.frame.}

\item{year_random}{A flag indicating whether to include year random effect. If FALSE, year is fitted as a fixed effect.}

\item{year_trend}{A flag indicating whether to fit a year trend effect.
Year trend cannot be fit if there is also a fixed year effect (as opposed to random effect).}

\item{adult_female_proportion}{A number between 0 and 1 of the expected proportion of adults that are female.
If NULL, a prior of dbeta(65, 35) is used. This prior can also be changed via the \code{priors} argument.}

\item{yearling_female_proportion}{A number between 0 and 1 of the expected proportion of yearlings that are female.}

\item{demographic_stochasticity}{A flag indicating whether to include demographic_stochasticity in the recruitment model.}

\item{priors}{A named vector of the parameter prior distribution values.
Any missing values are assigned their default values in \code{priors_survival()} and \code{priors_recruitment()}.
If NULL, all parameters are assigned their default priors.}
}
\description{
This is for use by developers.
}
\keyword{Internal}
